+++ grails-app/domain/de/laser/wekb/ElectronicBilling.groovy
@@ -1,9 +1,8 @@
-package de.laser
+package de.laser.wekb
 
+import de.laser.RefdataValue
 import de.laser.annotations.RefdataInfo
 import de.laser.storage.RDConstants
-import de.laser.wekb.Provider
-import de.laser.wekb.Vendor
 
 class ElectronicBilling implements Comparable<ElectronicBilling> {
