+++ grails-app/domain/de/laser/wekb/Vendor.groovy
@@ -3,10 +3,7 @@ package de.laser.wekb
 import de.laser.Address
 import de.laser.AlternativeName
 import de.laser.DocContext
-import de.laser.ElectronicBilling
-import de.laser.ElectronicDeliveryDelayNotification
 import de.laser.Identifier
-import de.laser.InvoiceDispatch
 import de.laser.LibrarySystem
 import de.laser.Org
 import de.laser.Person
