+++ grails-app/services/de/laser/GlobalSourceSyncService.groovy
@@ -14,6 +14,10 @@ import de.laser.storage.RDStore
 import de.laser.system.SystemEvent
 import de.laser.titles.TitleHistoryEvent
 import de.laser.utils.DateUtils
+import de.laser.wekb.ElectronicBilling
+import de.laser.wekb.ElectronicDeliveryDelayNotification
+import de.laser.wekb.InvoiceDispatch
+import de.laser.wekb.InvoicingVendor
 import de.laser.wekb.Package
 import de.laser.wekb.PackageVendor
 import de.laser.wekb.Platform
