+++ grails-app/services/de/laser/ProviderService.groovy
@@ -8,6 +8,9 @@ import de.laser.properties.ProviderProperty
 import de.laser.remote.ApiSource
 import de.laser.storage.RDStore
 import de.laser.traces.DeletedObject
+import de.laser.wekb.ElectronicBilling
+import de.laser.wekb.InvoiceDispatch
+import de.laser.wekb.InvoicingVendor
 import de.laser.wekb.Package
 import de.laser.wekb.Platform
 import de.laser.wekb.Provider
