+++ grails-app/services/de/laser/VendorService.groovy
@@ -11,6 +11,9 @@ import de.laser.survey.SurveyConfigVendor
 import de.laser.traces.DeletedObject
 import de.laser.utils.LocaleUtils
 import de.laser.utils.SwissKnife
+import de.laser.wekb.ElectronicBilling
+import de.laser.wekb.ElectronicDeliveryDelayNotification
+import de.laser.wekb.InvoiceDispatch
 import de.laser.wekb.Package
 import de.laser.wekb.PackageVendor
 import de.laser.wekb.Platform
